move operators and subject from examples to pushmi
