+++ examples/composition/composition_4.cpp
@@ -5,7 +5,7 @@
 
 #include <pool.h>
 
-#include <request_via.h>
+#include <pushmi/o/request_via.h>
 
 #include <pushmi/o/tap.h>
 #include <pushmi/o/transform.h>
@@ -14,7 +14,7 @@ using namespace pushmi::aliases;
 
 template<class Io>
 auto io_operation(Io io) {
-    return io | 
+    return io |
       op::transform([](auto){ return 42; }) |
       op::tap([](int v){ printf("io pool producing, %d\n", v); }) |
       op::request_via();
@@ -28,13 +28,13 @@ int main()
   auto io = ioPool.executor();
   auto cpu = cpuPool.executor();
 
-  io_operation(io).via([cpu]{ return cpu; }) | 
+  io_operation(io).via([cpu]{ return cpu; }) |
     op::tap([](int v){ printf("cpu pool processing, %d\n", v); }) |
     op::submit();
 
   // when the caller is not going to process the result (only side-effect matters)
   // or the caller is just going to push the result into a queue.
-  // provide a way to skip the transition to a different executor and make it 
+  // provide a way to skip the transition to a different executor and make it
   // stand out so that it has to be justified in code reviews.
   mi::via_cast<mi::is_sender<>>(io_operation(io)) | op::submit();
 
@@ -43,6 +43,3 @@ int main()
 
   std::cout << "OK" << std::endl;
 }
-
-
-
