+++ examples/then_execute/then_execute_2.cpp
@@ -13,12 +13,10 @@
 
 #include <pool.h>
 
-#include <request_via.h>
-#include <share.h>
-
 #include <pushmi/deferred.h>
 #include <pushmi/single_deferred.h>
 #include <pushmi/o/just.h>
+#include <pushmi/o/via.h>
 #include <pushmi/o/transform.h>
 
 using namespace pushmi::aliases;
@@ -38,8 +36,8 @@ public:
 };
 
 namespace p1054 {
-// A promise refers to a promise and is associated with a future, 
-// either through type-erasure or through construction of an 
+// A promise refers to a promise and is associated with a future,
+// either through type-erasure or through construction of an
 // underlying promise with an overload of make_promise_contract().
 
 // make_promise_contract() cannot be written to produce a lazy future.
@@ -103,4 +101,3 @@ int main()
 
   std::cout << "OK" << std::endl;
 }
-
