+++ include/pushmi/o/defer.h
@@ -6,9 +6,9 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-#include <pushmi/single.h>
-#include <pushmi/o/submit.h>
-#include <pushmi/o/extension_operators.h>
+#include "../single.h"
+#include "submit.h"
+#include "extension_operators.h"
 
 #if __cpp_deduction_guides >= 201703
 #define MAKE(x) x MAKE_
