+++ include/pushmi/o/just.h
@@ -7,6 +7,8 @@
 // LICENSE file in the root directory of this source tree.
 
 #include "../single_deferred.h"
+#include "submit.h"
+#include "extension_operators.h"
 
 namespace pushmi {
