+++ include/pushmi/o/share.h
@@ -6,11 +6,11 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-#include <pushmi/single.h>
-#include <pushmi/o/submit.h>
-#include <pushmi/o/extension_operators.h>
+#include "../single.h"
+#include "submit.h"
+#include "extension_operators.h"
 
-#include <subject.h>
+#include "../subject.h"
 
 namespace pushmi {
 
@@ -42,7 +42,7 @@ auto share_fn<T>::operator()() const {
 namespace operators {
 
 template<class T>
-PUSHMI_INLINE_VAR constexpr detail::share_fn<T> share{}; 
+PUSHMI_INLINE_VAR constexpr detail::share_fn<T> share{};
 
 } // namespace operators
